requirements.txt
discord.py
requests
nltk
opencv-python
pyyaml